backend/requirements.txt
fastapi>=0.104.1
websockets>=12.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
